add member ability to hold funds, add contribution approvals

+++ app/Http/Controllers/ContributionController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-use App\{Contribution, Company};
+use App\{Contribution, Company, Member};
 use Session;
 
 class ContributionController extends Controller
@@ -44,7 +44,9 @@ class ContributionController extends Controller
      */
     public function create()
     {
-        return view('pages.contribution_add');
+        $members = DB::table('members')->get();
+        return view('pages.contribution_add')
+            ->with('members',  $members);
     }
 
     /**
@@ -59,21 +61,15 @@ class ContributionController extends Controller
             'member_id' => 'required|exists:members,id',
             'period' => 'required|string',
             'amount' => 'required|numeric|min:1',
+            'fund_collector' => 'required|exists:members,id',
             'remarks' => 'sometimes|string|max:255',
         ]);
-
 
         DB::beginTransaction();
         try {
             $contribution = new Contribution;
             $contribution->fill($validated);
-            $contribution->search_text = "$contribution->member_id $contribution->period $contribution->amount $contribution->remarks";
             $contribution->save();
-
-            $company = Company::lockForUpdate()->firstOrFail();
-            $company->fund_total += $contribution->amount;
-            $company->fund_available += $contribution->amount;
-            $company->save();
 
             DB::commit();
         } catch (\Exception $e) {
@@ -108,9 +104,11 @@ class ContributionController extends Controller
         $contribution_id = $request['id'] ?? 0;
 
         $contribution = Contribution::findOrFail($contribution_id);
+        $members = DB::table('members')->get();
 
         return view('pages.contribution_edit')
-            ->with('contribution',  $contribution);
+            ->with('contribution',  $contribution)
+            ->with('members',  $members);
     }
 
     /**
@@ -131,15 +129,20 @@ class ContributionController extends Controller
             'remarks' => 'sometimes|string|max:255'
         ]);
 
-        
+        $contribution_id = $request['id'] ?? 0;
+        $contribution = Contribution::findOrFail($contribution_id);
+
+        if ($contribution->is_approved == true) {
+            Session::flash('error_message', "Contribution is already approved!");
+
+            return redirect()->route('contribution.edit', ['id' => $contribution_id]);
+        }
 
         DB::beginTransaction();
         try {
-            $contribution_id = $request['id'] ?? 0;
-            $contribution = Contribution::findOrFail($contribution_id);
+            
             // save the old contribution amount
             $old_contribution_amount = $contribution->amount;
-            
             $contribution->fill($validated);
             $contribution->search_text = "$contribution->member_id $contribution->period $contribution->amount $contribution->remarks";
             $contribution->save();
@@ -177,17 +180,25 @@ class ContributionController extends Controller
      */
     public function destroy(Request $request)
     {
+        // find the record and delete it
+        $contribution_id = $request['id'] ?? 0;
+        $contribution = Contribution::findOrFail($contribution_id);
+        
+        if ($contribution->is_approved == true) {
+            Session::flash('error_message', "Contribution is already approved!");
+
+            return redirect()->route('contribution.edit', ['id' => $contribution_id]);
+        }
+
         DB::beginTransaction();
         try {
-            // find the record and delete it
-            $contribution_id = $request['id'] ?? 0;
-            $contribution = Contribution::findOrFail($contribution_id);
+            
             $contribution->delete();
 
-            $company = Company::lockForUpdate()->firstOrFail();
-            $company->fund_total -= $contribution->amount;
-            $company->fund_available -= $contribution->amount;
-            $company->save();
+            // $company = Company::lockForUpdate()->firstOrFail();
+            // $company->fund_total -= $contribution->amount;
+            // $company->fund_available -= $contribution->amount;
+            // $company->save();
 
             DB::commit();
         } catch (\Exception $e) {
@@ -200,5 +211,59 @@ class ContributionController extends Controller
 
         // go back to the index page
         return redirect()->route('contribution.index');
+    }
+
+    /**
+     * Approve contribution
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function approve(Request $request)
+    {
+        // update the contribution
+        $contribution_id = $request['id'] ?? 0;
+        $contribution = Contribution::lockForUpdate()
+            ->findOrFail($contribution_id);
+
+        if ($contribution->is_approved == true) {
+            Session::flash('error_message', "Contribution is already approved!");
+
+            return redirect()->route('contribution.edit', ['id' => $contribution_id]);
+        }
+
+        DB::beginTransaction();
+        try {
+            
+            $contribution->is_approved = true;
+            $contribution->save();
+
+            // update members total contribution
+            $member = Member::lockForUpdate()
+                ->findOrFail($contribution->member_id);
+            $member->total_contribution += $contribution->amount;
+            $member->save();
+
+            // update collector's fund
+            $member = Member::lockForUpdate()
+                ->findOrFail($contribution->fund_collector);
+            $member->fund_on_hand += $contribution->amount;
+            $member->save();
+
+            $company = Company::lockForUpdate()->firstOrFail();
+            $company->fund_total += $contribution->amount;
+            $company->fund_available += $contribution->amount;
+            $company->save();
+
+            DB::commit();
+        } catch (\Exception $e) {
+            DB::rollBack();
+            throw $e;
+        }
+
+        // create success message 
+        Session::flash('success_message', "contribution ID [' $contribution_id '] has been approved!");
+
+        // go back to the index page
+        return redirect()->route('contribution.edit', ['id' => $contribution_id]);
     }
 }
